requirements copy.txt
Django @ git+https://github.com/stockitup/django.git@main
celery @ git+https://github.com/stockitup/celery.git@main
django-amazon-ses==4.0.0
django-celery-beat==2.2.0
django-crispy-forms==1.14.0
django-mathfilters==1.0.0
django-ses==2.0.0
djangorestframework==3.12.4
daphne==3.0.2
channels==3.0.5
channels-redis==3.4.1
numpy==1.24.4
psycopg2==2.8.6
PyMuPDF==1.23.19

requests==2.26.0
sendgrid==6.6.0
sentry-sdk==1.45.0
SQLAlchemy==1.4.6
service-identity==18.1.0
pyOpenSSL==24.0.0
Pillow==9.3.0
django-anymail==8.5
psutil==5.9.4
django-redis==5.2.0
django-rq==2.10.2
rq==1.16.1
setproctitle==1.2.2
crispy_bootstrap5==0.6
xlsxwriter==3.0.3
pandas==1.5.1
openpyxl==3.0.10
boto3==1.18.36
django-storages==1.13.1
beautifulsoup4==4.11.1
pywebpush==1.9.4
fpdf==1.7.2

pyinstrument==4.6.2

# spapi onzin
cachetools==5.2.1
confuse==2.0.0
pyjwt==2.6.0
pycryptodome==3.10.1
websockets==11.0.2
css-inline==0.11.0
python-barcode==0.15.1
xmltodict==0.12.0
pdfkit==1.0.0
cryptography==42.0.5
paramiko==3.4.0

#celery
billiard>=3.6.3.0,<4.0
kombu>=5.0.0,<6.0
vine>=5.0.0,<6.0
click>=7.0,<8.0
click-didyoumean>=0.0.3
click-repl>=0.1.6
click-plugins>=1.1.1

asyncpg==0.30.0
aiohttp==3.10.11
redis==5.2.0
uvicorn==0.32.1
jinja2==3.1.4